Moved material files to medium folder
